Add files via upload
Place these files in a directory respective to the number of teams you have.  Change the config.ini file for the respective team.  Setup the permissions on monitor.py with chmod 700 monitor.py.  Then execute as the root user ./monitor.py.  Execute each file for each respective team.  I am trying to keep it simple.

The csv file that is generated can be opened and adjusted in excel.  This is written in python 2.7 and has been tested on Debian/Ubuntu 64bit.  Should work on multiple platforms.